Move Environment.iOS.cs under libraries
The file does not contain anything Mono-specific and there is Environment.iOS.cs under libraries already.
