+++ src/mono/System.Private.CoreLib/System.Private.CoreLib.csproj
@@ -279,15 +279,6 @@
         <Link>Common\Interop\Unix\System.Native\Interop.GetEnviron.cs</Link>
       </Compile>
   </ItemGroup>
-  <ItemGroup Condition="'$(TargetsMacCatalyst)' == 'true' or '$(TargetsiOS)' == 'true' or '$(TargetstvOS)' == 'true'">
-      <Compile Include="$(BclSourcesRoot)\System\Environment.iOS.cs" />
-      <Compile Include="$(CommonPath)Interop\OSX\System.Native\Interop.SearchPath.cs">
-        <Link>Common\Interop\OSX\Interop.SearchPath.cs</Link>
-      </Compile>
-      <Compile Include="$(CommonPath)Interop\OSX\System.Native\Interop.SearchPath.iOS.cs">
-        <Link>Common\Interop\OSX\Interop.SearchPath.iOS.cs</Link>
-      </Compile>
-  </ItemGroup>
   <ItemGroup Condition="'$(TargetsBrowser)' == 'true'">
       <Compile Include="$(BclSourcesRoot)\System\Threading\TimerQueue.Browser.Mono.cs" />
   </ItemGroup>
